SPSE Module 4 Lesson 4. Ex 1 Create a PoC. SQL Injection using Mechanize This PoC is created taking BadStore as vulnerable app (192.168.1.4) Badstore official Web Page: http://badstore.net/
I found a XSS on the "Sign our Guestbook! " Link. That's what I will try to
automate (injecting XSS in forms). Link is: http://192.168.1.4/cgi-bin/badstore.cgi?action=guestbook

ToDo: This is just a PoC but it would be nice if I could follow all the links in a given HTML file
and test all the forms I could find per link followed.
Of course, this would be better if I could control the level of depth I want my program to go.